Move SellGUI to gui package
