+++ even-more-fish-plugin/src/main/java/com/oheers/fish/gui/SellGUI.java
@@ -1,4 +1,4 @@
-package com.oheers.fish.selling;
+package com.oheers.fish.gui;
 
 import com.github.Anon8281.universalScheduler.scheduling.tasks.MyScheduledTask;
 import com.oheers.fish.EvenMoreFish;
@@ -6,6 +6,8 @@ import com.oheers.fish.FishUtils;
 import com.oheers.fish.config.GUIConfig;
 import com.oheers.fish.config.GUIFillerConfig;
 import com.oheers.fish.config.MainConfig;
+import com.oheers.fish.selling.SellHelper;
+import com.oheers.fish.selling.WorthNBT;
 import com.oheers.fish.utils.GUIUtils;
 import de.themoep.inventorygui.GuiStorageElement;
 import de.themoep.inventorygui.InventoryGui;
